# swagger2.yaml
swagger: '2.0'
info:
  title: Adminiculum API
  description: Legal Workflow Platform API
  version: '1.0'
host: adminiculumaustriaeast-01.azurewebsites.net
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/v1/client-portal/summary/{clientId}:
    get:
      summary: GetClientSummary
      description: Ügyfél összefoglaló
      operationId: GetClientSummary
      produces:
        - application/json
      parameters:
        - name: clientId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Client summary
  /api/v1/client-portal/departments/{clientId}:
    get:
      summary: GetClientDepartments
      description: Ügyfél osztályai
      operationId: GetClientDepartments
      produces:
        - application/json
      parameters:
        - name: clientId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Client departments
  /api/v1/client-portal/departments/{deptId}/matters:
    get:
      summary: GetClientMatters
      description: Ügyfél ügyei
      operationId: GetClientMatters
      produces:
        - application/json
      parameters:
        - name: deptId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Client matters
  /api/v1/client-portal/matters/{matterId}:
    get:
      summary: GetClientMatterDetail
      description: Ügy részletei
      operationId: GetClientMatterDetail
      produces:
        - application/json
      parameters:
        - name: matterId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Matter details
  /api/v1/client-portal/matters/{matterId}/time-log:
    get:
      summary: GetClientTimeLog
      description: Ügy időnaplója
      operationId: GetClientTimeLog
      produces:
        - application/json
      parameters:
        - name: matterId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Time log
...
  /api/v1/matters/{id}:
    get:
      summary: GetMatterById
      description: Matter részletei
      operationId: GetMatterById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Matter details
...
securityDefinitions:
  oauth2-auth:
    type: oauth2
    flow: accessCode
    tokenUrl: https://login.microsoftonline.com/18b56834-dfea-4931-bdf8-e5ebb0cb4e0f/oauth2/v2.0/token
    authorizationUrl: https://login.microsoftonline.com/18b56834-dfea-4931-bdf8-e5ebb0cb4e0f/oauth2/v2.0/authorize
    scopes:
      api://a1e8b8a0-7690-4d09-9974-e4742d3de4e9/access_as_user: Access as user
security:
  - oauth2-auth:
      - api://a1e8b8a0-7690-4d09-9974-e4742d3de4e9/access_as_user